AddedTonsOfPowerUps
Added a bunch of power up beetles that are all functional.
The move speed stat may need to be fixed back to the controller using the characterdata, we'll see.

+++ Assets/Scripts/Pickups/HealthBeetle.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class HealthBeetle : MonoBehaviour, ICollectible
 {
-    public int healthToRestore;
+    public float healthToRestore;
 
     public void Collect()
     {
